Made first version of the giasmus editor.
The giasmus editor is a simple tkinter text-widget that allows the user to make and save texts with a chiastic structure.
The user has to input half of the chiasm, where each element of the chiasm needs to be on a separate line. Pressing the 'chiasmate' button will complete the chiasm.